review: use NUMERIC(12,2) for account.overdraft_limit
Aligns with translation-rules.md §2 which mandates BigDecimal/scale-2 SQL
for all money columns (overdraft limits explicitly listed). Matches the
existing available_balance / actual_balance shape on the same table.

+++ src/main/resources/db/migration/V0__baseline.sql
@@ -31,7 +31,7 @@ CREATE TABLE account (
     account_type         VARCHAR(8)     NOT NULL DEFAULT '',
     interest_rate        NUMERIC(6,2)   NOT NULL DEFAULT 0,
     opened               DATE           NOT NULL,
-    overdraft_limit      INTEGER        NOT NULL DEFAULT 0,
+    overdraft_limit      NUMERIC(12,2)  NOT NULL DEFAULT 0,
     last_stmt_date       DATE,
     next_stmt_date       DATE,
     available_balance    NUMERIC(12,2)  NOT NULL DEFAULT 0,
